refactor identity router
